refactor(runner): isolate lab patch apply
